Use context vars instead of terraform vars
