+++ amazon-msk/cdk/secure-public-access/lib/secure-public-access-stack.ts
@@ -57,11 +57,11 @@ export class ZillaPlusSecurePublicAccessStack extends cdk.Stack {
 
     const publicVar = zillaPlusContext.public;
     const mandatoryPublicVariables = [
-      'tlsCertificateKey',
+      'certificate',
       'wildcardDNS'
     ];
     validateContextKeys(publicVar, mandatoryPublicVariables);
-    const publicTlsCertificateKey = publicVar.tlsCertificateKey;
+    const publicTlsCertificateKey = publicVar.certificate;
     const publicWildcardDNS = publicVar.wildcardDNS;
 
     const vpc = ec2.Vpc.fromLookup(this, 'Vpc', { vpcId: vpcId });
@@ -393,7 +393,7 @@ systemctl start nitro-enclaves-acm.service
     data.public = {
       ...data.public,
       port: publicPort,
-      tlsCertificateKey: publicTlsCertificateKey,
+      certificate: publicTlsCertificateKey,
       wildcardDNS: publicWildcardDNS
     }
     data.externalHost = externalHost;
