+++ amazon-msk/cdk/web-streaming/lib/web-streaming-stack.ts
@@ -11,8 +11,7 @@ interface TemplateData {
   name: string;
   glue?: object;
   cloudwatch?: object;
-  path?: string;
-  topic?: string;
+  mappings?: Array<object>;
   public?: object;
   kafka?: object;
   jwt?: object;
@@ -27,7 +26,7 @@ export class WebStreamingStack extends cdk.Stack {
       'vpcId',
       'msk',
       'public',
-      'topic',
+      'mappings',
     ];
     
     function validateContextKeys(node: object, keys: string[]): void {
@@ -57,15 +56,20 @@ export class WebStreamingStack extends cdk.Stack {
     const mskCredentialsSecretName = msk.credentials;
     const publicVar = zillaPlusContext.public;
     const mandatoryPublicVariables = [
-      'tlsCertificateKey',
+      'certificate',
     ];
     validateContextKeys(publicVar, mandatoryPublicVariables);
 
-    const publicTlsCertificateKey = publicVar.tlsCertificateKey;
-    const kafkaTopic = zillaPlusContext.topic;
+    const publicTlsCertificateKey = publicVar.certificate;
+    const mappings = zillaPlusContext.mappings;
 
-    const customPath = zillaPlusContext.customPath;
-    const path = customPath ?? `/${kafkaTopic}`;
+    mappings.forEach((mapping: { path: string; topic: string; }) => {
+      if (!mapping.path) {
+        mapping.path = `/${mapping.topic}`;
+      }
+    });
+
+    const kafkaTopics: string[] = mappings.map((mapping: { topic: any; }) => mapping.topic);
 
     const vpc = ec2.Vpc.fromLookup(this, 'Vpc', { vpcId: vpcId });
     const subnets = vpc.selectSubnets();
@@ -343,10 +347,9 @@ export class WebStreamingStack extends cdk.Stack {
     }
     data.public = {
       port: publicPort,
-      tlsCertificateKey: publicTlsCertificateKey
+      certificate: publicTlsCertificateKey
     }
-    data.path = path;
-    data.topic = kafkaTopic;
+    data.mappings = mappings;
 
     const kafkaTopicCreationDisabled = zillaPlusContext.kafkaTopicCreationDisabled ?? false;
 
@@ -367,6 +370,12 @@ action=/opt/aws/bin/cfn-init -v --stack ${id} --resource ZillaPlusLaunchTemplate
 runas=root
     `;
 
+    let topicsCommand = "";
+    kafkaTopics.forEach((t: String) => {
+      topicsCommand = topicsCommand.concat(`
+./kafka-topics.sh --create --if-not-exists --bootstrap-server ${mskBootstrapServers} --command-config client.properties --replication-factor 2 --partitions 3 --topic ${t} --config 'cleanup.policy=compact'`);
+    });
+
     let kafkaTopicCreationCommand = "";
 
     if (!kafkaTopicCreationDisabled) {
@@ -385,7 +394,7 @@ sasl.jaas.config=org.apache.kafka.common.security.scram.ScramLoginModule require
 security.protocol=SASL_SSL
 sasl.mechanism=SCRAM-SHA-512
 EOF
-./kafka-topics.sh --create --if-not-exists --bootstrap-server ${mskBootstrapServers} --command-config client.properties --replication-factor 2 --partitions 3 --topic ${kafkaTopic} --config 'cleanup.policy=compact'
+${topicsCommand}
   `;
     }
