--- amazon-msk/cdktf/iot-ingest-and-control/.env.example
@@ -1,7 +1,0 @@
-MQTT_KAFKA_TOPIC_CREATION_DISABLED=false
-CREATE_ZILLA_PLUS_ROLE=true
-CREATE_ZILLA_PLUS_SECURITY_GROUP=true
-CLOUDWATCH_DISABLED=false
-SSH_KEY_ENABLED=false
-CLOUDWATCH_DISABLED=false
-
