+++ amazon-msk/cdktf/iot-ingest-and-control/cdktf.json
@@ -8,6 +8,17 @@
   ],
   "terraformModules": [],
   "context": {
-
+    "zilla-plus":
+    {
+      "msk":
+      {
+        "cluster": "my-msk-cluster",
+        "credentials": "AmazonMSK_alice"  
+      },
+      "public":
+      {
+        "certificate": "arn:aws:secretsmanager:us-east-1:445711703002:secret:wildcard.example.aklivity.io-9-u4J0YL"
+      }
+    }
   }
 }
