--- amazon-msk/cdktf/secure-public-access/.env.example
@@ -1,8 +1,0 @@
-CREATE_ZILLA_PLUS_ROLE=true
-CREATE_ZILLA_PLUS_SECURITY_GROUP=true
-SSH_KEY_ENABLED=false
-CLOUDWATCH_DISABLED=false
-#MSK_ACCESS_METHOD=<MSK access method override>
-
-## mTLS Specific Variables
-#PUBLIC_CERTIFICATE_AUTHORITY=
