--- amazon-msk/cdktf/secure-public-access/terraform.tfvars.example
@@ -1,14 +1,0 @@
-msk_cluster_name="<your msk cluster name>"
-public_tls_certificate_key="<your public tls certificate key ARN>"
-public_wildcard_dns="<your public wildcard dns>"
-
-## mTLS Specific Variables
-#msk_mtls_certificate_authority_arn="<your msk certificate authority arn>"
-
-## optional
-#zilla_plus_security_groups=["<your zilla plus security group>"]
-#zilla_plus_role="<your zilla plus role>"
-#public_mtls_certificate_authority_arn="<your public certificate authority arn>"
-#zilla_plus_ssh_key="<your ssh key name>"
-#cloudwatch_logs_group="<your cloudwatch logs group>"
-#cloudwatch_metrics_namespace="<your cloudwatch metrics namespace>"
