--- amazon-msk/cdktf/web-streaming/.env.example
@@ -1,8 +1,0 @@
-CREATE_ZILLA_PLUS_ROLE=true
-CREATE_ZILLA_PLUS_SECURITY_GROUP=true
-CLOUDWATCH_DISABLED=false
-GLUE_REGISTRY_ENABLED=false
-KAFKA_TOPIC_CREATION_DISABLED=false
-SSH_KEY_ENABLED=false
-CLOUDWATCH_DISABLED=false
-#CUSTOM_PATH=<you custom REST path>
