+++ amazon-msk/cdktf/web-streaming/package.json
@@ -22,7 +22,6 @@
   },
   "dependencies": {
     "@cdktf/provider-aws": "^19.27.0",
-    "@dotenvx/dotenvx": "^1.6.4",
     "cdktf": "^0.20.8",
     "constructs": "^10.3.0",
     "mustache": "^4.2.0"
