--- amazon-msk/cdktf/web-streaming/terraform.tfvars.example
@@ -1,16 +1,0 @@
-msk_cluster_name="<your msk cluster name>"
-msk_credentials_secret_name="<your msk credentials secret name>"
-kafka_topic="<your kafka topic>"
-public_tls_certificate_key="<your public tls certificate key ARN>"
-
-## optional
-#public_tcp_port=<your public tcp port>
-#zilla_plus_security_groups=["<your zilla plus security group>"]
-#zilla_plus_role="<your zilla plus role>"
-#zilla_plus_ssh_key="<your ssh key name>"
-#cloudwatch_logs_group="<your cloudwatch logs group>"
-#cloudwatch_metrics_namespace="<your cloudwatch metrics namespace>"
-#glue_registry="<your glue registry>"
-#jwt_issuer="<your jwt issuer>"
-#jwt_audience="<your jwt audience>"
-#jwt_keys_url="<jour jwks keys urls>"
