+++ amazon-msk/cdktf/web-streaming/tsconfig.json
